Make sure we load table data after start up.
Table previously uses a lazy load mechanism. This would make the initial
toggle being slow. This has been an issue on certain platform because
it's introducing a noticable sutter.

Unlikely pinyin, we need to load data based on the input method group,
so we need to defer until the input method manager is ready.