Remove the old string based trigger and replace everything with regex
If not talking about performance, feature-wise it is super set and the
thing we want to add has overlap with existing options (any "." will
enter quickphrase mode).

So just simply remove the old option seems to be a much clearer solution
and even if it will reset users existing configuration.